+++ app/frontend/stylesheets/object/component/_search-field-view.sass
@@ -1,6 +1,6 @@
 $SIZE: 30px
 
-.search-form-ivew
+.search-field-ivew
   > .searchform
     margin-bottom: 4px
     > .searchinput
