+++ app/frontend/stylesheets/object/component/_suggest-view.sass
@@ -7,6 +7,7 @@ $PADDING_SIDE: 14px
   display: flex
   font-size: 14px
   line-height: 24px
+  color: $COLOR_BLACK
   > .column
     + .column
       border-left: solid 1px $COLOR_SEPARATOR_SUB
