+++ app/frontend/stylesheets/object/project/_AdvancedSearchView.sass
@@ -1,3 +1,3 @@
 #AdvancedSearchView
-  > .search-form-ivew
+  > .search-field-ivew
     padding: 14px $PADDING_SIDE 0
