+++ app/frontend/stylesheets/object/project/_SimpleSearchView.sass
@@ -1,5 +1,5 @@
 #SimpleSearchView
-  > .search-form-ivew
+  > .search-field-ivew
     padding: 14px $PADDING_SIDE 14px
   > .suggest-view
     position: absolute
