+++ app/frontend/views/includes/_search_results_main.pug
@@ -7,7 +7,7 @@
   .sectionscontainer
 
     #SimpleSearchView.section
-      .search-form-ivew
+      .search-field-ivew
         .searchform
           input#search-field.searchinput(type="text" placeholder="Search for disease or gene symbol or rs...")
           button.searchbutton
@@ -32,7 +32,7 @@
       .suggest-view
 
     #AdvancedSearchView.section
-      //- .search-form-ivew
+      //- .search-field-ivew
       //-   .searchform
       //-     input#advanced-search-field.searchinput(type="text" placeholder="Search for disease or gene symbol or rs...")
       //-     button.searchbutton
